added optimied solution of triplet sum
